+++ examples/Hello/Hello.py
@@ -35,17 +35,15 @@ head = """
 """
 
 body = """
-<div style="display: table; margin: 50px auto auto auto;">
- <fieldset>
-  <input id="input" maxlength="20" placeholder="Enter a name here" type="text"
-         data-xdh-onevent="Submit" value="World"/>
-  <div style="display: flex; justify-content: space-around; margin: 5px auto auto auto;">
-   <button data-xdh-onevent="Submit">Submit</button>
-   <button data-xdh-onevent="Clear">Clear</button>
-   <button data-xdh-onevent="Error">Error</button>
-  </div>
- </fieldset>
+<fieldset>
+<input id="input" maxlength="20" placeholder="Enter a name here" type="text"
+        data-xdh-onevent="Submit" value="World"/>
+<div style="display: flex; justify-content: space-around; margin: 5px auto auto auto;">
+  <button data-xdh-onevent="Submit">Submit</button>
+  <button data-xdh-onevent="Clear">Clear</button>
+  <button data-xdh-onevent="Error">Error</button>
 </div>
+</fieldset>
 """
 
 def ac_connect(dom):